remove subprograms utils file
